remove packer
